\name{KetcheR-package}
\alias{KetcheR-package}
\alias{KetcheR}
\docType{package}
\title{
\packageTitle{KetcheR}
}
\description{
\packageDescription{KetcheR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KetcheR}
\packageIndices{KetcheR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{KetcheR}

Maintainer: \packageMaintainer{KetcheR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
